how to use useRef focus
